Port Python live-testing bug fixes into the C# rig/sync/websdr layers
Brings the C# port's already-completed Rig/Sync/WebSdr code up to date with
the v2.2.1-v2.2.4 bug-fix round shipped on master (merged in at b39d51c),
which had been ported from older Python source:

- flrig set_vfoA now sent as a double, not int (real flrig's XML-RPC
  signature is strictly "d:d"); added double support to XmlRpcCodec
- rigctld SetModeAsync always sends hamlib's -1 "leave bandwidth alone"
  sentinel instead of 0, so reverse-sync mode changes never touch the
  rig's filter/passband
- fixed the same stuck-candidate threshold bug as the Python side (forward
  and reverse FreqChangeThresholdHz both 0 now, was 10/50)
- added the rig-connect generation counter (engine-side half only, no C#
  GUI yet to consume it)
- websdr.org verify-readback now reads window.freq instead of the
  nonexistent #freqinput element

372/372 tests passing (up from 361). See PORT_BRIEF.md for full detail,
including a test-infrastructure race (shared CoreLog.Logger static +
xUnit cross-class parallelism) uncovered and fixed along the way.

+++ csharp/SDRSync.Rig/FlrigClient.cs
@@ -243,12 +243,20 @@ public sealed class FlrigClient : IRigClient, IAsyncDisposable
     /// Reverse-sync (WebSDR -> rig). set_vfoA's RPC response is not a
     /// reliable success signal, so this verifies via a bounded
     /// poll-until-match readback of GetFreqAsync() instead.
+    ///
+    /// Sends freqHz as a double, not an int: real flrig registers
+    /// rig.set_vfoA with XML-RPC signature "d:d" (confirmed against
+    /// flrig's own C++ source, src/server/xml_server.cxx, which does
+    /// `(double)params[0]`), and its bundled XmlRpc++ library enforces
+    /// that signature strictly by wire type -- sending a plain &lt;i4&gt;
+    /// serializes as the wrong type and real flrig rejects it outright
+    /// with Fault -1 "type error", before ever reaching the rig.
     /// </summary>
     public async Task<bool> SetFreqAsync(int freqHz, double? verifyBudgetS = null)
     {
         if (_http is null) return false;
         var budget = verifyBudgetS ?? SetVerifyBudgetS;
-        await CallAsync("rig.set_vfoA", new object?[] { freqHz });
+        await CallAsync("rig.set_vfoA", new object?[] { (double)freqHz });
 
         var deadline = DateTime.UtcNow + TimeSpan.FromSeconds(budget);
         while (true)
